Fix Import VM as template.
The ansible service broker sandbox cannot write a template into
the openshift namespace, therefore we need to ask for admin user
name and password for the import as template use case. In order
to do so we have added extra plans for import as template that
contain the username and passwod boxes.

Signed-off-by: Alexander Wels <awels@redhat.com>